Fixes to compile with DMD's new and improved import visibility/accessibility rules.
The unittest fixes are because of DMD issue #10378, and are only required because the
unittests are in the same file as the generic version of the templates. (Symbols in the same
file now take priority over non-selective imports, even if the import is done in a narrower
scope.)

The `contract.smartop` thing should never have worked; I believe the fix for DMD #314 is
what revealed my mistake there.

Older versions of `checkedint` included some deliberate abuse of DMD #313, but I think
I removed all that a while ago.